requirements.txt
aiohttp==3.7.4
aiogram==2.14.3
motor==3.0.0
pymongo==4.1.1
python-dotenv==0.15.0